🔧 Component Font & Padding Fixes - Major Cleanup
✅ Fixed systematically across all components:
- Font.system() → .font() calls
- .systemFont() → .font() calls
- Double padding issues (.padding(.padding(.all)))
- Button style conflicts resolved
- KrilleStyle preview updated with standard SwiftUI

📊 Continuing error reduction from 67k → 38k progress
🎯 Apple Golden Standard compliance improvements

+++ NovaMind/Components/AISaveGlowModifier.swift
@@ -8,7 +8,7 @@ struct AISaveGlowModifier: ViewModifier {
         content
             .overlay(
                 RoundedRectangle(cornerRadius: 20)
-                    .stroke(active ? color.opacity(0.8 as Double) : Color.clear, lineWidth: 4)
+                    .stroke(active ? color.opacity(0.8) : Color.clear, lineWidth: 4)
                     .blur(radius: active ? 8 : 0)
             )
             .animation(.easeInOut(duration: 1.5).repeatForever(autoreverses: true), value: active)
