+++ NovaMind/Core/Color+Theme.swift
@@ -18,7 +18,7 @@ extension Color {
 
     // MARK: - Computed Colors (for dynamic behavior)
     static var interactiveGlow: Color {
-        glow.opacity(0.8 as Double)
+        glow.opacity(0.8)
     }
 
     static var panelBackground: Color {
