+++ NovaMind/Realtime/WebSocketViews.swift
@@ -26,9 +26,9 @@ public struct WebSocketConnectionStatusView: View {
     .background(
       RoundedRectangle(cornerRadius: 6, style: .continuous)
         .fill(Color(NSColor.systemGray))
-        .shadow(color: .black.opacity(0.05 as Double), radius: 2, x: 0, y: 1)
+        .shadow(color: .black.opacity(0.05), radius: 2, x: 0, y: 1)
     )
-    .opacity(0.9 as Double)
+    .opacity(0.9)
   }
 
   private var statusText: String {
